Color inputs now with Text inputs some Refactoring and bugfixed
