+++ src/components/CardContainer/Cards/TestButton.jsx
@@ -1,0 +1,12 @@
+export default function TestButton({ theme, setTestThemeId }) {
+  return (
+    <button
+      className="test-button"
+      onClick={() => {
+        setTestThemeId(theme.id);
+      }}
+    >
+      Test Theme
+    </button>
+  );
+}
